Change Tender.contract_value to a BigIntegerField becuase some values were out of range.

+++ bidhive/bidhive_tendersearch/tender/models.py
@@ -18,7 +18,7 @@ class Tender(models.Model):
     uri = models.CharField(max_length=1024)
     # Where countries are sourced from via scrapy
     country = models.CharField(max_length=1024, null=True, choices=country_choices)
-    contract_value = models.PositiveIntegerField(null=True)
+    contract_value = models.BigIntegerField(null=True)
     contract_currency = models.CharField(max_length=3, null=True)
     description = models.TextField(null=True)
     status = models.CharField(max_length=1024, null=True)
